requirements.txt
langchain
langchain_core
langchain_groq
python_dotenv
langchain_community
langgraph
ipython
fastapi 
uvicorn
pydantic
